backend_or_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pydantic-ai-slim[openai,anthropic]>=0.8.1,<0.9
sqlmodel
psycopg[binary]
alembic
python-jose[cryptography]
passlib
